Update ember-cli, tether, and ember versions, and introduce bodyElementId option
We have updated to a new version of Tether that supports an option for specifying the element which Tether should move the positioned element to (the default is the body). This is particularly useful for Ember acceptance tests, which root the Ember app in the `#ember-testing` div.

```js
// config/environment.js

ENV['ember-tether'] = {
  bodyElementId: 'ember-testing'
};
```

It is also possible to pass a `bodyElement` to a particular `ember-tether` component declaration.