Fix failing test in ember-lts-2.12
This specific version of Ember tests fail with `You have turned on
testing mode, which disabled the run-loop's autorun. You will need to
wrap any code with asynchronous side-effects in a run` when the tooltip
is re-rendered. Wrapping the changes which kick off the render in a run
loop solves this issue.